refactor: move component from nested folders
