+++ src/shared/components/footer.tsx
