+++ src/shared/components/index.ts
@@ -1,11 +1,8 @@
 export * from './ui/accordion'
 export * from './ui/button'
-export * from './ui/toast'
 export * from './ui/tooltip'
-
-export * from './common/content'
-export * from './common/centered-layout'
-export * from './common/footer'
-export * from './common/header'
-
-export * from './animation/transition'
+export * from './content'
+export * from './centered-layout'
+export * from './footer'
+export * from './header'
+export * from './transition'
